+++ src/ADSDREnvelope.cpp
@@ -134,7 +134,6 @@ struct ADSDREnvelopeModule : InfNoiseEnvelopeModule {
         configLight(CLIP_RANGE_LIGHT, getClipRangeLightName(outClipRange.act));
 
         envelope = releaseLevel;
-        prevEnvelope = envelope;
         configInput(PHASE_INPUT, "Phase (Gate/Trigger)");
         configSwitch(PHASE_GATE_TRIG_PARAM, 0.f, 1.f, 1.f, "Phase-mode", {"Trigger", "Gate"});
         configInput(A_TRIG_INPUT, "Attack/Decay Trigger");
@@ -194,6 +193,11 @@ struct ADSDREnvelopeModule : InfNoiseEnvelopeModule {
         ensureNormExpLogLuts();
 	}
 
+    void onTrigLengthChanged() override {
+        for (int i = 0; i < 6; i++)
+            outTrig[i].setCycles(trigOnOffCycles);
+    }
+
     void onReset(const ResetEvent& e) override {
         InfNoiseEnvelopeModule::onReset(e);
 
@@ -213,8 +217,6 @@ struct ADSDREnvelopeModule : InfNoiseEnvelopeModule {
             outTrig[i].reset();
         phaseTrigMode = false;
         envelope = releaseLevel;
-        prevEnvelope = envelope;
-        envMotion = em_steady;
         rampFrom = releaseLevel;
         rampTo = attackLevel;
         oldAttackShape = 2.f;
@@ -442,8 +444,6 @@ struct ADSDREnvelopeModule : InfNoiseEnvelopeModule {
             ((cycle256 & processQualityPatterns[procQuality.act]) == processQualityPatterns[procQuality.act]));
 
         if (doProcess) {
-            prevEnvelope = envelope;
-
             // Handle direct A.trig
             float attVolt = haveAttTrigInput ? inputs[A_TRIG_INPUT].getVoltage() : 0.f;
             if (params[A_TRIG_BTN_PARAM].getValue() > 0.5f)
@@ -502,12 +502,12 @@ struct ADSDREnvelopeModule : InfNoiseEnvelopeModule {
             }
 
             // Process trigger outputs
-            outTrig[BOA].process(procSampleTime);
-            outTrig[EOA].process(procSampleTime);
-            outTrig[BOS].process(procSampleTime);
-            outTrig[EOS].process(procSampleTime);
-            outTrig[BOR].process(procSampleTime);
-            outTrig[EOR].process(procSampleTime);
+            outTrig[BOA].process(procCycles);
+            outTrig[EOA].process(procCycles);
+            outTrig[BOS].process(procCycles);
+            outTrig[EOS].process(procCycles);
+            outTrig[BOR].process(procCycles);
+            outTrig[EOR].process(procCycles);
     
             // Handle triggered phase change/retrig or transition/delay/hold
             if (attTriggered) {
@@ -580,7 +580,6 @@ struct ADSDREnvelopeModule : InfNoiseEnvelopeModule {
                 outputs[ENVELOPE_OUTPUT].setVoltage(envelope);
                 outputs[INV_ENVELOPE_OUTPUT].setVoltage(attackLevel + releaseLevel - envelope);
             }
-            updateEnvMotion();
             pushToExpanders();
         }